+++ packages/devtools_app/lib/src/http/http_request_data.dart
@@ -12,7 +12,6 @@ import 'package:vm_service/vm_service.dart';
 import '../config_specific/logger/logger.dart';
 import '../primitives/utils.dart';
 import '../screens/network/network_model.dart';
-import '../shared/common_widgets.dart';
 import '../shared/globals.dart';
 import 'http.dart';
 
@@ -264,11 +263,6 @@ class DartIOHttpRequestData extends NetworkRequest {
       if (!_request.isResponseComplete) return null;
       if (_responseBody != null) return _responseBody;
       _responseBody = utf8.decode(fullRequest.responseBody!);
-      if (contentType != null && contentType!.contains('json')) {
-        _responseBody = FormattedJson.encoder.convert(
-          json.decode(_responseBody!),
-        );
-      }
       return _responseBody;
     } on FormatException {
       return '<binary data>';
