+++ packages/devtools_app/lib/src/screens/debugger/variables.dart
@@ -5,19 +5,20 @@
 import 'package:flutter/material.dart' hide Stack;
 import 'package:flutter/rendering.dart';
 import 'package:provider/provider.dart';
+import 'package:vm_service/vm_service.dart';
 
 import '../../primitives/listenable.dart';
 import '../../primitives/utils.dart';
 import '../../shared/common_widgets.dart';
 import '../../shared/globals.dart';
 import '../../shared/notifications.dart';
+import '../../shared/object_tree.dart';
 import '../../shared/routing.dart';
 import '../../shared/theme.dart';
 import '../../shared/tree.dart';
 import '../inspector/diagnostics.dart';
 import '../inspector/inspector_screen.dart';
 import 'debugger_controller.dart';
-import 'debugger_model.dart';
 
 class Variables extends StatelessWidget {
   const Variables({Key? key}) : super(key: key);
@@ -100,7 +101,7 @@ Widget displayProvider(
       TextSpan(
         children: processAnsiTerminalCodes(
           variable.text,
-          theme.fixedFontStyle,
+          theme.subtleFixedFontStyle,
         ),
       ),
       onTap: onTap,
@@ -115,17 +116,48 @@ Widget displayProvider(
       debuggerController: controller,
     );
   }
+  TextStyle variableDisplayStyle() {
+    final style = theme.subtleFixedFontStyle;
+    switch (variable.ref!.instanceRef!.kind) {
+      case InstanceKind.kString:
+        return style.apply(
+          color: theme.colorScheme.stringSyntaxColor,
+        );
+      case InstanceKind.kInt:
+        return style.apply(
+          color: theme.colorScheme.numericConstantSyntaxColor,
+        );
+      case InstanceKind.kBool:
+      case InstanceKind.kNull:
+        return style.apply(
+          color: theme.colorScheme.modifierSyntaxColor,
+        );
+      default:
+        return style;
+    }
+  }
+
+  final hasName = variable.name?.isNotEmpty ?? false;
   return DevToolsTooltip(
     message: variable.displayValue,
     waitDuration: tooltipWaitLong,
     child: SelectableText.rich(
       TextSpan(
-        text: variable.name?.isNotEmpty ?? false ? '${variable.name}: ' : null,
-        style: theme.fixedFontStyle,
+        text: hasName ? variable.name : null,
+        style: variable.artificialName
+            ? theme.subtleFixedFontStyle
+            : theme.fixedFontStyle.apply(
+                color: theme.colorScheme.controlFlowSyntaxColor,
+              ),
         children: [
+          if (hasName)
+            TextSpan(
+              text: ': ',
+              style: theme.fixedFontStyle,
+            ),
           TextSpan(
             text: variable.displayValue,
-            style: theme.subtleFixedFontStyle,
+            style: variableDisplayStyle(),
           ),
         ],
       ),
