+++ packages/devtools_app/lib/src/screens/inspector/diagnostics.dart
@@ -6,11 +6,11 @@ import 'package:flutter/material.dart';
 
 import '../../primitives/utils.dart';
 import '../../shared/globals.dart';
+import '../../shared/object_tree.dart';
 import '../../shared/theme.dart';
 import '../../ui/hover.dart';
 import '../../ui/icons.dart';
 import '../debugger/debugger_controller.dart';
-import '../debugger/debugger_model.dart';
 import '../debugger/variables.dart';
 import 'diagnostics_node.dart';
 import 'inspector_controller.dart';
