+++ packages/devtools_app/lib/src/screens/network/network_request_inspector_views.dart
@@ -149,9 +149,7 @@ class HttpRequestView extends StatelessWidget {
     return Padding(
       padding: const EdgeInsets.all(denseSpacing),
       child: SingleChildScrollView(
-        child: FormattedJson(
-          formattedString: data.requestBody,
-        ),
+        child: JsonViewer(encodedJson: data.requestBody!),
       ),
     );
   }
@@ -165,22 +163,24 @@ class HttpResponseView extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     Widget child;
+    final theme = Theme.of(context);
     // We shouldn't try and display an image response view when using the
     // timeline profiler since it's possible for response body data to get
     // dropped.
     final contentType = data.contentType;
+    final responseBody = data.responseBody!;
     if (contentType != null && contentType.contains('image')) {
       child = ImageResponseView(data);
+    } else if (contentType != null &&
+        contentType.contains('json') &&
+        responseBody.isNotEmpty) {
+      child = JsonViewer(encodedJson: responseBody);
     } else {
-      child = FormattedJson(
-        formattedString: data.responseBody,
+      child = Text(
+        responseBody,
+        style: theme.fixedFontStyle,
       );
     }
-    // TODO(kenz): use a collapsible json tree for the response
-    // https://github.com/flutter/devtools/issues/2952.
-    // TODO(bkonyi): add syntax highlighting to these responses
-    // https://github.com/flutter/devtools/issues/2604. We may be able to use
-    // the new tree widget used in the Provider page.
     return Padding(
       padding: const EdgeInsets.all(denseSpacing),
       child: SingleChildScrollView(child: child),
