+++ packages/devtools_app/lib/src/service/vm_service_wrapper.dart
@@ -10,6 +10,7 @@ import 'package:flutter/foundation.dart';
 import 'package:vm_service/vm_service.dart';
 
 import '../primitives/utils.dart';
+import 'json_to_service_cache.dart';
 
 class VmServiceWrapper implements VmService {
   VmServiceWrapper(
@@ -71,6 +72,10 @@ class VmServiceWrapper implements VmService {
     ..complete(true);
 
   Future<void> get allFuturesCompleted => _allFuturesCompleter.future;
+
+  // A local cache of "fake" service objects. Used to convert JSON objects to
+  // VM service response formats to be used with APIs that require them.
+  final fakeServiceCache = JsonToServiceCache();
 
   /// Executes `callback` for each isolate, and waiting for all callbacks to
   /// finish before completing.
@@ -284,6 +289,14 @@ class VmServiceWrapper implements VmService {
     int? offset,
     int? count,
   }) {
+    final cachedObj = fakeServiceCache.getObject(
+      objectId: objectId,
+      offset: offset,
+      count: count,
+    );
+    if (cachedObj != null) {
+      return Future.value(cachedObj);
+    }
     return trackFuture(
       'getObject',
       _vmService.getObject(
