+++ packages/devtools_app/lib/src/shared/console_service.dart
@@ -12,6 +12,7 @@ import '../screens/inspector/diagnostics_node.dart';
 import '../screens/inspector/inspector_service.dart';
 import '../service/vm_service_wrapper.dart';
 import 'globals.dart';
+import 'object_tree.dart';
 
 /// A line in the console.
 ///
