+++ packages/devtools_app/test/debugger/association_variable_test.dart
@@ -2,9 +2,9 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:devtools_app/src/screens/debugger/debugger_model.dart';
 import 'package:devtools_app/src/service/service_manager.dart';
 import 'package:devtools_app/src/shared/globals.dart';
+import 'package:devtools_app/src/shared/object_tree.dart';
 import 'package:devtools_test/devtools_test.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
@@ -79,13 +79,8 @@ void main() {
 
     await buildVariablesTree(variable);
 
-    expect(variable.children.first.children, [
-      matchesVariable(name: '[key]', value: '\'Hey\''),
-      matchesVariable(name: '[value]', value: '12.34'),
-    ]);
-
     expect(variable.children, [
-      matchesVariable(name: null, value: '[Entry 0]'),
+      matchesVariable(name: 'Hey', value: '12.34'),
     ]);
   });
 
@@ -140,11 +135,7 @@ void main() {
     await buildVariablesTree(variable);
 
     expect(variable.children, [
-      matchesVariable(name: null, value: '[Entry 0]'),
-    ]);
-    expect(variable.children.first.children, [
-      matchesVariable(name: '[key]', value: '1'),
-      matchesVariable(name: '[value]', value: '12.34'),
+      matchesVariable(name: '1', value: '12.34'),
     ]);
   });
